Overhaul Control Panel permissions system, using policy methods

+++ app/controllers/admin/elements_controller.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 module Admin
 
   class ElementsController < ControlPanelController
@@ -5,21 +7,22 @@ module Admin
     include ActionController::Live
 
     class ImportMode
-      MERGE = 'merge'
+      MERGE   = 'merge'
       REPLACE = 'replace'
     end
 
     PERMITTED_PARAMS = [:description, :name]
 
-    before_action :authorize_modify_elements, only: [:create, :destroy, :edit,
-                                                     :import, :update]
     before_action :set_permitted_params
+    before_action :set_element, except: [:create, :import, :index]
+    before_action :authorize_element, except: [:create, :import, :index]
 
     ##
     # XHR only
     #
     def create
       @element = Element.new(sanitized_params)
+      authorize(@element)
       begin
         @element.save!
       rescue ActiveRecord::RecordInvalid
@@ -39,65 +42,57 @@ module Admin
     end
 
     def destroy
-      element = Element.find_by_name(params[:name])
-      raise ActiveRecord::RecordNotFound unless element
-      begin
-        element.destroy!
-      rescue => e
-        handle_error(e)
-      else
-        flash['success'] = "Element \"#{element.name}\" deleted."
-      ensure
-        redirect_back fallback_location: admin_elements_path
-      end
+      @element.destroy!
+    rescue => e
+      handle_error(e)
+    else
+      flash['success'] = "Element \"#{@element.name}\" deleted."
+    ensure
+      redirect_back fallback_location: admin_elements_path
     end
 
     ##
     # XHR only
     #
     def edit
-      element = Element.find_by_name(params[:name])
-      raise ActiveRecord::RecordNotFound unless element
-
       render partial: 'admin/elements/form',
-             locals: { element: element }
+             locals: { element: @element }
     end
 
     ##
     # Responds to POST /admin/elements/import
     #
     def import
-      begin
-        raise 'No elements specified.' if params[:elements].blank?
-
-        json = params[:elements].read.force_encoding('UTF-8')
-        struct = JSON.parse(json)
-        ActiveRecord::Base.transaction do
-          if params[:import_mode] == ImportMode::REPLACE
-            Element.delete_all # skip callbacks & validation
-          end
-          struct.each do |hash|
-            e = Element.find_by_name(hash['name'])
-            if e
-              e.update_from_json_struct(hash)
-            else
-              Element.from_json_struct(hash).save!
-            end
+      authorize(Element)
+      raise 'No elements specified.' if params[:elements].blank?
+      json   = params[:elements].read.force_encoding('UTF-8')
+      struct = JSON.parse(json)
+      ActiveRecord::Base.transaction do
+        if params[:import_mode] == ImportMode::REPLACE
+          Element.delete_all # skip callbacks & validation
+        end
+        struct.each do |hash|
+          e = Element.find_by_name(hash['name'])
+          if e
+            e.update_from_json_struct(hash)
+          else
+            Element.from_json_struct(hash).save!
           end
         end
-      rescue => e
-        handle_error(e)
-        redirect_to admin_elements_path
-      else
-        flash['success'] = "#{struct.length} elements created or updated."
-        redirect_to admin_elements_path
       end
+    rescue => e
+      handle_error(e)
+      redirect_to admin_elements_path
+    else
+      flash['success'] = "#{struct.length} elements created or updated."
+      redirect_to admin_elements_path
     end
 
     ##
     # Responds to GET /elements
     #
     def index
+      authorize(Element)
       respond_to do |format|
         format.html do
           sql = 'SELECT elements.id, elements.name, elements.description,
@@ -110,7 +105,7 @@ module Admin
               FROM elements
               GROUP BY elements.id, elements.name
               ORDER BY elements.name;'
-          @elements = ActiveRecord::Base.connection.exec_query(sql)
+          @elements    = ActiveRecord::Base.connection.exec_query(sql)
           @new_element = Element.new
         end
         format.json do
@@ -125,32 +120,26 @@ module Admin
     # Responds to GET /admin/elements/:name
     #
     def show
-      @element = Element.find_by_name(params[:name])
-      raise ActiveRecord::RecordNotFound unless @element
     end
 
     ##
     # XHR only
     #
     def update
-      element = Element.find_by_name(params[:name])
-      raise ActiveRecord::RecordNotFound unless element
-      begin
-        element.update!(sanitized_params)
-      rescue ActiveRecord::RecordInvalid
-        response.headers['X-Kumquat-Result'] = 'error'
-        render partial: 'shared/validation_messages',
-               locals: { entity: element }
-      rescue => e
-        handle_error(e)
-        keep_flash
-        render 'admin/shared/reload'
-      else
-        response.headers['X-Kumquat-Result'] = 'success'
-        flash['success'] = "Element \"#{element.name}\" updated."
-        keep_flash
-        render 'admin/shared/reload'
-      end
+      @element.update!(sanitized_params)
+    rescue ActiveRecord::RecordInvalid
+      response.headers['X-Kumquat-Result'] = 'error'
+      render partial: 'shared/validation_messages',
+             locals:  { entity: @element }
+    rescue => e
+      handle_error(e)
+      keep_flash
+      render 'admin/shared/reload'
+    else
+      response.headers['X-Kumquat-Result'] = 'success'
+      flash['success'] = "Element \"#{@element.name}\" updated."
+      keep_flash
+      render 'admin/shared/reload'
     end
 
     ##
@@ -158,15 +147,12 @@ module Admin
     # all usages of a given element by all items.
     #
     def usages
-      element = Element.find_by_name(params[:element_name])
-      raise ActiveRecord::RecordNotFound unless element
-
-      response.headers['Content-Type'] = 'text/plain'
-      response.headers['Content-Disposition'] = "attachment;filename=#{element.name}.tsv"
+      response.headers['Content-Type']        = 'text/tab-separated-values; charset=utf-8'
+      response.headers['Content-Disposition'] = "attachment;filename=#{@element.name}.tsv"
 
       response.stream.write "collection_id\titem_id\telement_name\telement_value\telement_uri" +
                 ItemTsvExporter::LINE_BREAK
-      element.usages.each do |row|
+      @element.usages.each do |row|
         response.stream.write row.values.join("\t")
         response.stream.write ItemTsvExporter::LINE_BREAK
       end
@@ -174,17 +160,20 @@ module Admin
       response.stream.close
     end
 
+
     private
 
-    def authorize_modify_elements
-      unless current_user.can?(Permissions::MODIFY_ELEMENTS)
-        flash['error'] = 'You do not have permission to perform this action.'
-        redirect_to admin_elements_url
-      end
+    def authorize_element
+      @element ? authorize(@element) : skip_authorization
     end
 
     def sanitized_params
       params.require(:element).permit(PERMITTED_PARAMS)
+    end
+
+    def set_element
+      @element = Element.find_by_name(params[:name] || params[:element_name])
+      raise ActiveRecord::RecordNotFound unless @element
     end
 
     def set_permitted_params
